Build complete APK: fix all compilation errors and move MobSpawner package
- Rewrote switch expressions/multi-case labels to Java 11 syntax (d8 compat)
- Added SANDSTONE to BlockType enum
- Moved MobSpawner to entity.mob package for package-private mob access
- Fixed NoiseGenerator float cast, HUD switch, SoundManager switch
- Removed duplicate mob classes from RemainingMobs.java
- Compiled with Java 11 javac to avoid MethodParameters null bug in d8
- Built signed debug APK: minecraft2.apk (103KB)

+++ Minecraft2Android/app/src/main/java/com/minecraft2/android/world/NoiseGenerator.java
@@ -48,7 +48,7 @@ public class NoiseGenerator {
     public double octaveNoise(float x, float y, int octaves, float persistence) {
         double total = 0, freq = 1, amp = 1, maxVal = 0;
         for (int i = 0; i < octaves; i++) {
-            total += noise(x * freq, y * freq) * amp;
+            total += noise((float)(x * freq), (float)(y * freq)) * amp;
             maxVal += amp;
             amp *= persistence;
             freq *= 2;
